backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymupdf==1.23.8
sentence-transformers==2.2.2
chromadb==0.4.22
google-generativeai==0.3.2
numpy==1.24.3
pydantic==2.5.0
httpx==0.25.2
python-jose==3.3.0
aiofiles==23.2.1
Pillow==10.1.0
huggingface-hub==0.16.4
torch==2.0.1
transformers==4.30.2